# Performance improvements for Encoding & Decoding