--- a/src/Sqids.Benchmarks/Program.cs
+++ b/src/Sqids.Benchmarks/Program.cs
@@ -1,0 +1,5 @@
+﻿using BenchmarkDotNet.Running;
+
+BenchmarkSwitcher
+	.FromAssembly(typeof(Program).Assembly)
+	.RunAllJoined();