@marengohue 2025-05-13
It would be great if the encoder supported Span<T> or Memory<T> as an input to allow for buffer encoding with minimal allocation. It wouldn’t necessarily improve the performance of Sqids, but rather allow the consumers to invoke it with fewer allocations.